# Frontend/quran-dashboard-ui/openapi/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "المنهج القرآني API",
    "version": "v1"
  },
  "paths": {
...
    "/api/abwab/protection/{categoryId}": {
      "get": {
        "tags": [
          "Protection"
        ],
        "parameters": [
          {
            "name": "categoryId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryProtectionProfileDtoApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryProtectionProfileDtoApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryProtectionProfileDtoApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/abwab/protection/{categoryId}/{protectionType}/apply": {
      "post": {
        "tags": [
          "Protection"
        ],
        "parameters": [
          {
            "name": "categoryId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "protectionType",
            "in": "path",
            "required": true,
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/ManualProtectionType"
                }
              ]
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/ApplyManualProtectionRequest"
                  }
                ]
              }
            },
            "text/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/ApplyManualProtectionRequest"
                  }
                ]
              }
            },
            "application/*+json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/ApplyManualProtectionRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/abwab/protection/{categoryId}/{protectionType}/lift": {
      "post": {
        "tags": [
          "Protection"
        ],
        "parameters": [
          {
            "name": "categoryId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "protectionType",
            "in": "path",
            "required": true,
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/ManualProtectionType"
                }
              ]
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/LiftManualProtectionRequest"
                  }
                ]
              }
            },
            "text/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/LiftManualProtectionRequest"
                  }
                ]
              }
            },
            "application/*+json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/LiftManualProtectionRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/abwab/protection/{categoryId}/full-preset": {
      "post": {
        "tags": [
          "Protection"
        ],
        "parameters": [
          {
            "name": "categoryId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/ApplyFullProtectionPresetRequest"
                  }
                ]
              }
            },
            "text/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/ApplyFullProtectionPresetRequest"
                  }
                ]
              }
            },
            "application/*+json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/ApplyFullProtectionPresetRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/abwab/sections": {
      "post": {
        "tags": [
          "Sections"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/AddSectionRequest"
                  }
                ]
              }
            },
            "text/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/AddSectionRequest"
                  }
                ]
              }
            },
            "application/*+json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/AddSectionRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/abwab/sections/{sectionId}": {
      "put": {
        "tags": [
          "Sections"
        ],
        "parameters": [
          {
            "name": "sectionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/EditSectionRequest"
                  }
                ]
              }
            },
            "text/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/EditSectionRequest"
                  }
                ]
              }
            },
            "application/*+json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/EditSectionRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Sections"
        ],
        "parameters": [
          {
            "name": "sectionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/DeleteSectionRequest"
                  }
                ]
              }
            },
            "text/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/DeleteSectionRequest"
                  }
                ]
              }
            },
            "application/*+json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/DeleteSectionRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/abwab/sections/reorder": {
      "post": {
        "tags": [
          "Sections"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/ReorderSectionsRequest"
                  }
                ]
              }
            },
            "text/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/ReorderSectionsRequest"
                  }
                ]
              }
            },
            "application/*+json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/ReorderSectionsRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AddSectionRequest": {
        "required": [
          "expectedTimelineGeneration",
          "expectedTreeRevision",
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "expectedTreeRevision": {
            "type": "integer",
            "format": "int64"
          },
          "expectedTimelineGeneration": {
            "type": "integer",
            "format": "int64"
          }
        },
        "additionalProperties": false
      },
...
      "ApplyFullProtectionPresetRequest": {
        "required": [
          "expectedTimelineGeneration",
          "expectedVersions",
          "scope"
        ],
        "type": "object",
        "properties": {
          "scope": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ManualProtectionScope"
              }
            ]
          },
          "expectedVersions": {
            "required": [
              "CategoryData",
              "InternalStructure",
              "QuranContent",
              "Deletion",
              "Relationship"
            ],
            "type": "object",
            "properties": {
              "CategoryData": {
                "type": "integer",
                "format": "int32"
              },
              "InternalStructure": {
                "type": "integer",
                "format": "int32"
              },
              "QuranContent": {
                "type": "integer",
                "format": "int32"
              },
              "Deletion": {
                "type": "integer",
                "format": "int32"
              },
              "Relationship": {
                "type": "integer",
                "format": "int32"
              }
            },
            "additionalProperties": false
          },
          "expectedTimelineGeneration": {
            "type": "integer",
            "format": "int64"
          }
        },
        "additionalProperties": false
      },
      "ApplyManualProtectionRequest": {
        "required": [
          "expectedTimelineGeneration",
          "expectedVersion",
          "scope"
        ],
        "type": "object",
        "properties": {
          "scope": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ManualProtectionScope"
              }
            ]
          },
          "expectedVersion": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "expectedTimelineGeneration": {
            "type": "integer",
            "format": "int64"
          }
        },
        "additionalProperties": false
      },
...
      "CategoryProtectionProfileDto": {
        "required": [
          "categoryId",
          "expectedTimelineGeneration",
          "manualProtections",
          "ordinaryProtection",
          "serverTimeUtc"
        ],
        "type": "object",
        "properties": {
          "expectedTimelineGeneration": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ExpectedTimelineGeneration"
              }
            ]
          },
          "categoryId": {
            "type": "string",
            "format": "uuid"
          },
          "manualProtections": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ManualProtectionResolutionDto"
            }
          },
          "ordinaryProtection": {
            "allOf": [
              {
                "$ref": "#/components/schemas/OrdinaryProtectionResolutionDto"
              }
            ]
          },
          "serverTimeUtc": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "CategoryProtectionProfileDtoApiResponse": {
        "required": [
          "data",
          "errors",
          "isSuccess",
          "message"
        ],
        "type": "object",
        "properties": {
          "isSuccess": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "data": {
            "allOf": [
              {
                "$ref": "#/components/schemas/CategoryProtectionProfileDto"
              }
            ],
            "nullable": true
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "DeleteSectionRequest": {
        "required": [
          "expectedTimelineGeneration",
          "expectedTreeRevision",
          "expectedVersion"
        ],
        "type": "object",
        "properties": {
          "expectedVersion": {
            "type": "integer",
            "format": "int32"
          },
          "expectedTreeRevision": {
            "type": "integer",
            "format": "int64"
          },
          "expectedTimelineGeneration": {
            "type": "integer",
            "format": "int64"
          }
        },
        "additionalProperties": false
      },
...
      "EditSectionRequest": {
        "required": [
          "expectedTimelineGeneration",
          "expectedTreeRevision",
          "expectedVersion",
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "expectedVersion": {
            "type": "integer",
            "format": "int32"
          },
          "expectedTreeRevision": {
            "type": "integer",
            "format": "int64"
          },
          "expectedTimelineGeneration": {
            "type": "integer",
            "format": "int64"
          }
        },
        "additionalProperties": false
      },
      "ExpectedTimelineGeneration": {
        "required": [
          "generation"
        ],
        "type": "object",
        "properties": {
          "generation": {
            "type": "integer",
            "format": "int64"
          }
        },
        "additionalProperties": false
      },
...
      "LiftManualProtectionRequest": {
        "required": [
          "expectedTimelineGeneration",
          "expectedVersion"
        ],
        "type": "object",
        "properties": {
          "expectedVersion": {
            "type": "integer",
            "format": "int32"
          },
          "expectedTimelineGeneration": {
            "type": "integer",
            "format": "int64"
          }
        },
        "additionalProperties": false
      },
...
      "ManualProtectionResolutionDto": {
        "required": [
          "actionClassification",
          "isDirect",
          "isProtected",
          "protectionType",
          "scope",
          "serverTimeUtc",
          "sourceCategoryId"
        ],
        "type": "object",
        "properties": {
          "protectionType": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ManualProtectionType"
              }
            ]
          },
          "isProtected": {
            "type": "boolean"
          },
          "isDirect": {
            "type": "boolean"
          },
          "sourceCategoryId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "scope": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ManualProtectionScope"
              }
            ],
            "nullable": true
          },
          "actionClassification": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ProtectionActionClassification"
              }
            ]
          },
          "serverTimeUtc": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "ManualProtectionScope": {
        "enum": [
          0,
          1
        ],
        "type": "integer",
        "format": "int32"
      },
      "ManualProtectionType": {
        "enum": [
          0,
          1,
          2,
          3,
          4
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "ObjectApiResponse": {
        "required": [
          "data",
          "errors",
          "isSuccess",
          "message"
        ],
        "type": "object",
        "properties": {
          "isSuccess": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "data": {
            "nullable": true
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "OrdinaryProtectionResolutionDto": {
        "required": [
          "actorSubject",
          "expiresAtUtc",
          "isActive",
          "lastEditedAtUtc"
        ],
        "type": "object",
        "properties": {
          "isActive": {
            "type": "boolean"
          },
          "actorSubject": {
            "type": "string",
            "nullable": true
          },
          "lastEditedAtUtc": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "expiresAtUtc": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ProtectionActionClassification": {
        "enum": [
          0,
          1,
          2
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "ReorderSectionsRequest": {
        "required": [
          "expectedTimelineGeneration",
          "expectedTreeRevision",
          "orders"
        ],
        "type": "object",
        "properties": {
          "orders": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SectionOrderEntryRequest"
            }
          },
          "expectedTreeRevision": {
            "type": "integer",
            "format": "int64"
          },
          "expectedTimelineGeneration": {
            "type": "integer",
            "format": "int64"
          }
        },
        "additionalProperties": false
      },
...
      "SectionOrderEntryRequest": {
        "required": [
          "expectedVersion",
          "sectionId",
          "sortOrder"
        ],
        "type": "object",
        "properties": {
          "sectionId": {
            "type": "string",
            "format": "uuid"
          },
          "sortOrder": {
            "type": "integer",
            "format": "int32"
          },
          "expectedVersion": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
    }
  },
...
}